feat: auth layout
