--- app/(auth)/_layout.tsx
@@ -1,11 +1,0 @@
-
-import { View, Text } from 'react-native'
-import React from 'react'
-import { Tabs } from "expo-router"
-
-// This tab is the index.tsx that is transform into tabs when partnered with layout supported by expo
-const AuthLayout = () => {
-	return <Tabs screenOptions={{headerShown: false, tabBarStyle: {display: "none"}}}/>
-}
-
-export default AuthLayout
