+++ app/_layout.tsx
@@ -5,10 +5,7 @@ import { Stack } from "expo-router"
 
 const AppLayout = () => {
   return (
-	<Stack screenOptions={{headerShown: false}} >
-		<Stack.Screen name="(auth)" />
-		<Stack.Screen name="(main)" />
-	</Stack>
+	<Stack screenOptions={{headerShown: false}} />
   )
 }
